fixed errors in: conditionFunctions, StateEventTypes, StateLighting, Stats, and UnitModeGroups

+++ dlls/Excel/Stats.cs
@@ -17,30 +17,30 @@ namespace Reanimator.Excel
             [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 64)]
             public string stat;
 
-            public Int16 code;
-            public Int16 type;
-            public Int16 assocStat1;
-            public Int16 assocStat2;
-            public Int16 regenIntervalInMS;
-            public Int16 regenDivisor;
-            public Int16 regenDelayOnDec;
-            public Int16 regenDelayOnZero;
-            public Int16 regenDelayMonster;
+            public Int32 code;
+            public Int32 type;
+            public Int32 assocStat1;
+            public Int32 assocStat2;
+            public Int32 regenIntervalInMS;
+            public Int32 regenDivisor;
+            public Int32 regenDelayOnDec;
+            public Int32 regenDelayOnZero;
+            public Int32 regenDelayMonster;
             [MarshalAs(UnmanagedType.ByValArray, SizeConst = 8)]
             byte[] undefined1;
-            public Int16 offset;
-            public Int16 shift;
-            public Int16 minSet;
-            public Int16 maxSet;
+            public Int32 offset;
+            public Int32 shift;
+            public Int32 minSet;
+            public Int32 maxSet;
             [MarshalAs(UnmanagedType.ByValArray, SizeConst = 4)]
             byte[] undefined3;
-            public Int16 minAssert;
-            public Int16 maxAssert;
-            public Int16 undefined4;
-            public Int16 accrueToTypes1;
-            public Int16 accrueToTypes2;
-            public Int16 unitType;
-            public Int16 bitMask1;/*1 bit cur
+            public Int32 minAssert;
+            public Int32 maxAssert;
+            public Int32 undefined4;
+            public Int32 accrueToTypes1;
+            public Int32 accrueToTypes2;
+            public Int32 unitType;
+            public Int32 bitMask1;/*1 bit cur
 	2 bit modlist
 	3 bit vector
 	4 bit float
@@ -61,55 +61,55 @@ namespace Reanimator.Excel
 	21 bit update database
 	22 bit don't transfer to nonweapon missile */
             [MarshalAs(UnmanagedType.ByValArray, SizeConst = 4)]
-            byte[] undefined5;
-            public Int16 player;//bool
-            public Int16 monster;
-            public Int16 missile;
-            public Int16 item;
-            public Int16 Object;
-            public Int16 valbits;
-            public Int16 valWindow;
+                        byte[] undefined5;
+            public Int32 player;//bool
+            public Int32 monster;
+            public Int32 missile;
+            public Int32 item;
+            public Int32 Object;
+            public Int32 valbits;
+            public Int32 valWindow;
             [MarshalAs(UnmanagedType.ByValArray, SizeConst = 4)]
             byte[] undefined6;
-            public Int16 valShift;
-            public Int16 valOffs;
-            public Int16 valTable;
+            public Int32 valShift;
+            public Int32 valOffs;
+            public Int32 valTable;
             [MarshalAs(UnmanagedType.ByValArray, SizeConst = 8)]
             byte[] undefined7;
-            public Int16 param1Bits;
-            public Int16 param2Bits;
-            public Int16 param3Bits;
-            public Int16 param4Bits;
+            public Int32 param1Bits;
+            public Int32 param2Bits;
+            public Int32 param3Bits;
+            public Int32 param4Bits;
             [MarshalAs(UnmanagedType.ByValArray, SizeConst = 16)]
             byte[] undefined8;
-            public Int16 param1Shift;
-            public Int16 param2Shift;
-            public Int16 param3Shift;
-            public Int16 param4Shift;
-            public Int16 param1Offs;
-            public Int16 param2Offs;
-            public Int16 param3Offs;
-            public Int16 param4Offs;
-            public Int16 param1Table;
-            public Int16 param2Table;
-            public Int16 param3Table;
-            public Int16 param4Table;
+            public Int32 param1Shift;
+            public Int32 param2Shift;
+            public Int32 param3Shift;
+            public Int32 param4Shift;
+            public Int32 param1Offs;
+            public Int32 param2Offs;
+            public Int32 param3Offs;
+            public Int32 param4Offs;
+            public Int32 param1Table;
+            public Int32 param2Table;
+            public Int32 param3Table;
+            public Int32 param4Table;
             [MarshalAs(UnmanagedType.ByValArray, SizeConst = 32)]
             byte[] undefined9;
-            public Int16 minTicksBetweenDbCommand;
-            public Int16 databaseUnitField;
-            public Int16 specFunc;
-            public Int16 sfStat1;
-            public Int16 sfStat2;
-            public Int16 sfStat3;
-            public Int16 sfStat4;
+            public Int32 minTicksBetweenDbCommand;
+            public Int32 databaseUnitField;
+            public Int32 specFunc;
+            public Int32 sfStat1;
+            public Int32 sfStat2;
+            public Int32 sfStat3;
+            public Int32 sfStat4;
             [MarshalAs(UnmanagedType.ByValArray, SizeConst = 28)]
             byte[] undefined10;
-            public Int16 checkAgainstTypes1;
-            public Int16 checkAgainstTypes2;
-            public Int16 reqFailString;//stridx
-            public Int16 reqFailStringHellgate;
-            public Int16 reqFailStringMythos;
+            public Int32 checkAgainstTypes1;
+            public Int32 checkAgainstTypes2;
+            public Int32 reqFailString;//stridx
+            public Int32 reqFailStringHellgate;
+            public Int32 reqFailStringMythos;
             [MarshalAs(UnmanagedType.ByValArray, SizeConst = 16)]
             byte[] undefined11;
             [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 64)]
